[Behat] Extract the ManagingPlacedOrderAddressesContext class
