+++ src/Sylius/Behat/Resources/config/services/contexts/api/admin.xml
@@ -275,5 +275,11 @@
             <argument type="service" id="Sylius\Behat\Client\ResponseCheckerInterface" />
             <argument type="service" id="api_platform.iri_converter" />
         </service>
+
+        <service id="Sylius\Behat\Context\Api\Admin\ManagingPlacedOrderAddressesContext">
+            <argument type="service" id="sylius.behat.api_platform_client.admin" />
+            <argument type="service" id="Sylius\Behat\Client\ResponseCheckerInterface" />
+            <argument type="service" id="sylius.behat.shared_storage" />
+        </service>
     </services>
 </container>
