+++ src/Sylius/Behat/Resources/config/suites.yml
@@ -28,7 +28,7 @@ imports:
     - suites/api/locale/locales.yaml
     - suites/api/locale/managing_locales.yml
     - suites/api/order/managing_orders.yml
-    - suites/api/order/modifying_address.yaml
+    - suites/api/order/modifying_placed_order_address.yaml
     - suites/api/payment/managing_payment_methods.yaml
     - suites/api/payment/managing_payments.yml
     - suites/api/product/adding_product_review.yml
